+++ src/components/landing/Button.tsx
@@ -1,7 +1,6 @@
-import React from "react";
 import { FaArrowRightLong } from "react-icons/fa6";
 
-const Button = (props) => {
+const Button = (props: any) => {
   return (
     <button className="px-8 py-3.5 bg-blue-d1 text-white rounded-[40px] font-semibold text-base flex items-center  gap-1 font-heading self-center">
       {props.text}
