+++ src/components/landing/JoinUsBanner.tsx
@@ -1,4 +1,3 @@
-import React from "react";
 import ban from "../../assets/landing/banner_image.png";
 const JoinUsBanner = () => {
   return (
