--- src/components/landing/Menu.jsx
@@ -1,22 +1,0 @@
-import React from "React";
-import "../../styles/landing/menu.css";
-import MenuCategories from "./MenuCategories";
-import MenuPosts from "./MenuPosts";
-
-const Menu = () => {
-  return (
-    <div className={"menu_container"}>
-      <h2 className={"subtitle"}>{"What's hot"}</h2>
-      <h1 className={"title"}>Most Popular</h1>
-      <MenuPosts withImage={false} />
-      <h2 className={"subtitle"}>Discover by topic</h2>
-      <h1 className={"title"}>Categories</h1>
-      <MenuCategories />
-      <h2 className={"subtitle"}>Chosen by the editor</h2>
-      <h1 className={"title"}>Editors Pick</h1>
-      <MenuPosts withImage={true} />
-    </div>
-  );
-};
-
-export default Menu;
