+++ src/pages/Home/sections/Home/Animation.tsx
@@ -1,4 +1,3 @@
-import React from "react";
 import AnimatedSVG from "../../../../components/landing/AnimatedSVG";
 import MobileAnimation from "../../../../components/landing/MobileAnimation";
 import TabletAnimation from "../../../../components/landing/TabletAnimation";
