--- src/pages/Home/sections/Home/Benefits.jsx
@@ -1,11 +1,0 @@
-import BenefitsCarousel from "../../../../components/landing/BenefitsCarousel";
-
-const Benefits = () => {
-  return (
-    <section className="flex items-center justify-center p-[100px]">
-      <BenefitsCarousel />
-    </section>
-  );
-};
-
-export default Benefits;
